Fix build system
The libssrckdtree package did not seem to build correctly. The files
necessary for building it are now in the lib/ folder. The base package's
pom.xml file was also updated to use a higher version Java compiler
(since certain pieces of the code require it). Finally, a text file
BUILDING was added to help explain how one builds the package.